Added Delete Student from DB using UI

+++ DayCare/src/edu/neu/csye6200/api/concrete/ConcreteStudentApi.java
@@ -62,7 +62,7 @@ public class ConcreteStudentApi implements StudentApi {
     }
 
     @Override
-    public void deleteStudent(int studentId) {
+    public void deleteStudent(long studentId) {
         StudentDao.deleteStudentDao(studentId);
     }
